requirements.txt
wheel
pyTelegramBotAPI
mysql-connector-python
